requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0

# YOLOv8 for detection
ultralytics>=8.0.0

# EasyOCR for license plate recognition
easyocr>=1.7.0

# DeepSORT for object tracking
deep-sort-realtime>=1.3.2

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
websockets>=11.0

# Configuration
pyyaml>=6.0

# Image processing
Pillow>=10.0.0
